refactor: move NewPatrimonyModel to domain layer
