+++ backend/src/data/protocols/db/patrimony/db-create-new-patrimony.ts
@@ -1,9 +1,4 @@
-import { Patrimony, PatrimonyItens } from '@/domain/model/patrimony'
-
-export interface NewPatrimonyModel extends Omit<Patrimony, 'id' | 'patrimonyItens'>{
-  patrimonyItens: Array<Omit<PatrimonyItens, 'id'>>
-}
-
+import { NewPatrimonyModel } from '@/domain/usecase/patrimony/create-patrimony'
 export interface DbCreateNewPatrimony {
   create(newPatrimony: NewPatrimonyModel): Promise<string>
 }
